Remove exceptions from the public interface for now
