+++ projects/core/src/main/resources/data/computercraft/lua/rom/apis/parallel.lua
@@ -37,8 +37,6 @@ parallel.waitForAny(do_sleep(), rednet.receive)
 @since 1.2
 ]]
 
-local exception
-
 local function create(...)
     local tFns = table.pack(...)
     local tCos = {}
@@ -68,8 +66,7 @@ local function runUntilLimit(_routines, _limit)
                 if tFilters[r] == nil or tFilters[r] == eventData[1] or eventData[1] == "terminate" then
                     local ok, param = coroutine.resume(r, table.unpack(eventData, 1, eventData.n))
                     if not ok then
-                        if exception == nil then exception = dofile("rom/modules/main/cc/exception.lua") end
-                        error(exception.wrap_error(param, r), 0)
+                        error(param, 0)
                     else
                         tFilters[r] = param
                     end
